API module renamed to RPC
